Extracts update logic to smart abstract class
